+++ cpp/jni/javet_monitor.cpp
@@ -29,8 +29,12 @@ namespace Javet {
 		}
 
 		jlongArray JavetNativeMonitor::GetCounters(JNIEnv* jniEnv) {
+			jlong buffer[CounterType::Max];
+			for (int i = 0; i < CounterType::Max; ++i) {
+				buffer[i] = counters[i].load();
+			}
 			jlongArray returnDataArray = jniEnv->NewLongArray(CounterType::Max);
-			jniEnv->SetLongArrayRegion(returnDataArray, 0, CounterType::Max, counters);
+			jniEnv->SetLongArrayRegion(returnDataArray, 0, CounterType::Max, buffer);
 			return returnDataArray;
 		}
 	}